requirements.txt
pandas
kagglehub
sqlite3
python-dotenv